fix: harden integer overflow in buffer sizing; document headers
Cast w*h to size_t before multiplying in lt__buffer_resize, lt_clear, and
the EGC sweep: the operands are int (up to ~65535 from ws_col/ws_row), so
the product could exceed INT_MAX and invoke signed-overflow UB. The result
was caught by the existing size guard in practice, but the multiply itself
was undefined.

Document the >=5-byte output contract for lt_utf8_unicode_to_char (up to 4
encoded bytes plus the NUL), which was previously unstated.

Add doc comments to every previously-undocumented function across the
public API (libterm.h) and the internal/platform headers, matching the
codebase's existing prose-comment style. Fix the stale lt_print comment
that claimed no wcwidth support — lt_print_ex already advances by
lt_wcwidth (wide=2, combining=0).

+++ src/shared/egc.c
@@ -61,13 +61,14 @@ static void lt__egc_sweep(void) {
   if (!live)
     return; /* can't sweep this round; correctness is unaffected, only memory */
 
-  const int n = lt__g.width * lt__g.height;
+  /* Cast before multiplying: int w * int h can overflow INT_MAX (UB). */
+  const size_t n = (size_t)lt__g.width * (size_t)lt__g.height;
   const struct lt_cell *bufs[2] = {lt__g.back, lt__g.front};
   for (int b = 0; b < 2; b++) {
     const struct lt_cell *buf = bufs[b];
     if (!buf)
       continue;
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
       uint32_t id = buf[i]._reserved;
       if (id > 0 && id < lt__egc_count)
         live[id] = true;
